eval follow-ups: address PR #737 review
- Surfaced interrupts emit resolvedBy null: nothing RESOLVED anything,
  and "handler" claimed a handler was involved even for an empty chain.
  Which handler propagated/passed is on the handlerDecision events. New
  test pins the empty-chain case (nobody was watching this effect).
- The empty-suite guard moves into loadInputs itself, covering
  eval optimize (which would have burned its whole iteration budget
  grading candidates against nothing) and every other caller.
- readEvalRun warns once when a run is all-missing but has records at
  the pre-#733 flat path, so grading an old run says "old layout" instead
  of looking like an agent that produced nothing.
- NormalizedEnvelope carries seq (statelog file position); the interrupt
  sort tiebreaks on it instead of relying on Array#sort stability for
  same-millisecond events.
- findRecursively -> lib/utils/findRecursively.ts and getImports/
  getAllImports -> lib/analysis/imports.ts, so lib/compiler (and the
  compile entry points) genuinely stop importing from lib/cli. The two
  remaining library-to-CLI edges are the node/agent execution helpers
  (agencyRunner -> runAgencyNode, judge/pairwise -> runAgencyAgent);
  cli/util.ts is two concepts wearing one name, split left for its own
  change.
