+++ packages/agency-lang/lib/eval/readRun.ts
@@ -46,7 +46,26 @@ export function readEvalRun(runDir: string): ReadEvalRunResult {
     };
   }
 
+  warnIfLegacyLayout(resolvedRunDir, inputsById);
   return { runDir: resolvedRunDir, inputsById };
+}
+
+/** The pre-#733 flat layout is no longer read. Without this, grading such a
+ *  run prints objective 0.00 with every input "missing" — which reads like an
+ *  agent that produced nothing, not like an old directory. Name what
+ *  happened, once. */
+function warnIfLegacyLayout(runDir: string, inputsById: Record<string, ReadEvalRunInput>): void {
+  const all = Object.values(inputsById);
+  const allMissing = all.length > 0 && all.every((input) => input.status === "missing");
+  if (!allMissing) return;
+  const hasFlatRecord = all.some((input) =>
+    fs.existsSync(path.join(runDir, "inputs", input.inputId, "eval-record.json")));
+  if (hasFlatRecord) {
+    console.warn(
+      `readEvalRun: ${runDir} uses the pre-#733 run layout and is no longer readable. ` +
+      `Re-run the suite, or read this directory with a build from before PR #737.`,
+    );
+  }
 }
 
 function inputStatus(result: EvalRunInputResult, recordPath: string): ReadEvalRunInput["status"] {
