+++ packages/agency-lang/lib/statelogClient.ts
@@ -986,7 +986,9 @@ export class StatelogClient {
   }: {
     interruptId: string;
     outcome: "approved" | "rejected" | "propagated" | "passed";
-    resolvedBy: "handler" | "user" | "policy" | "ipc";
+    /** Null when nothing resolved anything — the chain-outcome event a
+     *  surfacing interrupt records ("propagated"/"passed"). */
+    resolvedBy: "handler" | "user" | "policy" | "ipc" | null;
     timeTaken?: number;
     /** Optional summary of the interrupt being resolved. See
      *  `handlerDecision.interrupt` for rationale. */
